Refactor the base library
