+++ README.md
@@ -25,19 +25,20 @@ A Python toolkit for chain-of-thought prompting
 
 Cogitator is a Python toolkit for experimenting and working with
 [chain-of-thought (CoT) prompting](https://arxiv.org/abs/2201.11903)
-techniques in large language models (LLMs).
+methods in large language models (LLMs).
 CoT prompting improves LLM performance on complex tasks (like question-answering, reasoning, and problem-solving)
 by guiding the models to generate intermediate reasoning steps before arriving at the final answer.
-The toolkit aims to make it easier to try out popular CoT methods and frameworks for research or integrate them into AI
+Additionally, it can be used to improve the interpretability of LLMs by providing insight into the model's reasoning process.
+The toolkit aims to make it easier to use popular CoT strategies and frameworks for research or integrating them into AI
 applications.
 
 ### Features
 
-- Provides unified sync/async API for CoT methods
+- Provides unified sync/async API for CoT strategies
 - Supports using OpenAI and Ollama as LLM providers
 - Supports structured JSON model output with Pydantic validation
 - Includes a customizable benchmarking framework (see [benches](benches))
-- Includes implementations of popular CoT methods and frameworks like
+- Includes implementations of popular CoT strategies and frameworks like
     - [Self-Consistency CoT (ICLR 2023)](https://arxiv.org/abs/2203.11171)
     - [Automatic CoT (ICLR 2023)](https://arxiv.org/abs/2210.03493)
     - [Least-to-Most Prompting (ICLR 2023)](https://arxiv.org/abs/2205.10625)
@@ -87,9 +88,9 @@ except Exception as e:
     print("Please make sure Ollama is running and the model is pulled.")
     exit(1)
 
-# Step 3: Choose a CoT method (Self-Consistency in this case)
+# Step 3: Choose a CoT strategies (Self-Consistency in this case)
 # Self-Consistency generates multiple reasoning paths and finds the most common answer
-sc_method = SelfConsistency(
+sc_strategy = SelfConsistency(
     llm,
     n_samples=5,  # Number of reasoning paths to generate
     temperature=0.7  # Higher temperature can lead to more diverse answers
@@ -99,10 +100,10 @@ sc_method = SelfConsistency(
 question = "A bat and a ball cost $1.10 in total. The bat costs $1.00 more than the ball. How much does the ball cost?"
 prompt = f"Q: {question}\nA: Let's think step by step."
 
-# Step 5: Run the CoT prompting method
+# Step 5: Run the CoT prompting sc_strategy
 print(f"\nQuestion: {question}")
 print("Running Self-Consistency CoT...")
-final_answer = sc_method.run(prompt)  # Returns the most consistent (repeated) answer
+final_answer = sc_strategy.run(prompt)  # Returns the most consistent (repeated) answer
 
 # Expected output: $0.05 or 0.05 (may vary slightly based on model and temperature)
 print(f"\nCogitator's Answer (Self-Consistency): {final_answer}")
@@ -119,7 +120,8 @@ See the [docs](docs) directory for the documentation for the Cogitator toolkit.
 ### Benchmarking Framework
 
 This project includes a customizable and extensible benchmarking framework to evaluate the performance of different
-CoT methods on various datasets like [GSM8K](https://arxiv.org/abs/2110.14168) and [StrategyQA](https://arxiv.org/abs/2101.02235).
+CoT strategies on various datasets like [GSM8K](https://arxiv.org/abs/2110.14168) and
+[StrategyQA](https://arxiv.org/abs/2101.02235).
 
 Check out the [benches](benches) directory for more details about the framework and how it could be used.
 
@@ -131,8 +133,7 @@ See [CONTRIBUTING.md](CONTRIBUTING.md) for details on how to make a contribution
 
 ### Logo
 
-The logo is named "Cognition" and was originally created by
-[vectordoodle](https://www.svgrepo.com/author/vectordoodle).
+The logo is named "Cognition" and was originally created by [vectordoodle](https://www.svgrepo.com/author/vectordoodle).
 
 ### License
