+++ cogitator/__init__.py
@@ -1,21 +1,21 @@
 import importlib
 import logging
 
-from .auto_cot import AutoCoT
-from .cdw_cot import CDWCoT
 from .clustering import BaseClusterer, KMeansClusterer
 from .embedding import BaseEmbedder, SentenceTransformerEmbedder
-from .graph_of_thoughts import GraphOfThoughts
-from .least_to_most import LeastToMost
 from .model import BaseLLM, OllamaLLM, OpenAILLM
-from .sc_cot import SelfConsistency
 from .schemas import (
     EvaluationResult,
     ExtractedAnswer,
     LTMDecomposition,
     ThoughtExpansion,
 )
-from .tree_of_thoughts import TreeOfThoughts
+from .strategies import AutoCoT
+from .strategies import CDWCoT
+from .strategies import GraphOfThoughts
+from .strategies import LeastToMost
+from .strategies import SelfConsistency
+from .strategies import TreeOfThoughts
 from .utils import accuracy, approx_token_length, count_steps, exact_match
 
 _logger = logging.getLogger(__name__)
@@ -29,19 +29,15 @@ except importlib.metadata.PackageNotFoundError:
     )
 
 __all__ = [
-    # Chain of Thought methods and frameworks
     "AutoCoT",
     "BaseClusterer",
-    # Embedding and Clustering
     "BaseEmbedder",
-    # Model abstractions
     "BaseLLM",
     "CDWCoT",
     "EvaluationResult",
     "ExtractedAnswer",
     "GraphOfThoughts",
     "KMeansClusterer",
-    # Schemas
     "LTMDecomposition",
     "LeastToMost",
     "OllamaLLM",
@@ -52,7 +48,6 @@ __all__ = [
     "TreeOfThoughts",
     "accuracy",
     "approx_token_length",
-    # Utils (bunch of useful stuff)
     "count_steps",
     "exact_match",
 ]
