+++ cogitator/model/__init__.py
@@ -1,0 +1,9 @@
+from .base import BaseLLM
+from .ollama import OllamaLLM
+from .openai import OpenAILLM
+
+__all__ = [
+    "BaseLLM",
+    "OllamaLLM",
+    "OpenAILLM",
+]
