+++ docs/api/clustering.md
@@ -1,0 +1,16 @@
+# cogitator.clustering
+
+This module defines abstractions and implementations for clustering algorithms, primarily used to group similar embeddings in
+strategies like AutoCoT and CDWCoT.
+
+::: cogitator.clustering.BaseClusterer
+options:
+show_root_heading: false
+show_source: true
+members_order: source
+
+::: cogitator.clustering.KMeansClusterer
+options:
+show_root_heading: false
+show_source: true
+members_order: source
