+++ docs/api/embedding.md
@@ -1,0 +1,16 @@
+# cogitator.embedding
+
+This module provides abstractions and implementations for text embedding models, used by strategies like AutoCoT and CDWCoT to
+represent questions or thoughts numerically.
+
+::: cogitator.embedding.BaseEmbedder
+options:
+show_root_heading: false
+show_source: true
+members_order: source
+
+::: cogitator.embedding.SentenceTransformerEmbedder
+options:
+show_root_heading: false
+show_source: true
+members_order: source
