+++ docs/api/least_to_most.md
@@ -1,8 +1,8 @@
-# cogitator.least_to_most
+# cogitator.strategies.least_to_most
 
 Implementation of the Least-to-Most (LtM) prompting strategy.
 
-::: cogitator.least_to_most
+::: cogitator.strategies.least_to_most.LeastToMost
 options:
 show_root_heading: true
 show_source: true
