+++ docs/api/model.md
@@ -1,7 +1,9 @@
-# cogitator.model
+# cogitator.model.base
 
-::: cogitator.model
+This module defines the abstract base class for all LLM providers.
+
+::: cogitator.model.base.BaseLLM
 options:
 show_root_heading: true
 show_source: true
-# Add members_order, show_bases, etc. as needed
+members_order: source
