+++ docs/api/ollama.md
@@ -1,0 +1,9 @@
+# cogitator.model.ollama
+
+Implementation of the BaseLLM interface for Ollama models.
+
+::: cogitator.model.ollama.OllamaLLM
+options:
+show_root_heading: true
+show_source: true
+members_order: source
