+++ docs/api/openai.md
@@ -1,0 +1,9 @@
+# cogitator.model.openai
+
+Implementation of the BaseLLM interface for OpenAI models.
+
+::: cogitator.model.openai.OpenAILLM
+options:
+show_root_heading: true
+show_source: true
+members_order: source
