+++ docs/api/sc_cot.md
@@ -1,8 +1,8 @@
-# cogitator.sc_cot
+# cogitator.strategies.sc_cot
 
 Implementation of the Self-Consistency (SC-CoT) prompting strategy.
 
-::: cogitator.sc_cot
+::: cogitator.strategies.sc_cot.SelfConsistency
 options:
 show_root_heading: true
 show_source: true
