+++ docs/api/tree_of_thoughts.md
@@ -1,8 +1,8 @@
-# cogitator.tree_of_thoughts
+# cogitator.strategies.tree_of_thoughts
 
 Implementation of the Tree of Thoughts (ToT) prompting framework.
 
-::: cogitator.tree_of_thoughts
+::: cogitator.strategies.tree_of_thoughts.TreeOfThoughts
 options:
 show_root_heading: true
 show_source: true
