+++ docs/benchmarking.md
@@ -7,12 +7,12 @@ on various datasets.
 
 The framework consists of two main scripts:
 
-1.  **`benches/run.py`**: Handles the generation phase. It loads datasets, configures LLMs and CoT strategies based on
-    `benches.yml` and CLI arguments, runs the strategies on dataset questions, and saves the raw LLM outputs and timings to
-    a JSONL file.
-2.  **`benches/evaluate.py`**: Handles the evaluation phase. It reads the results file generated by `run.py`, extracts
-    final answers from the raw outputs (using either heuristics or an LLM), compares them to gold answers, and calculates
-    accuracy metrics for each method.
+1. **`benches/run.py`**: Handles the generation phase. It loads datasets, configures LLMs and CoT strategies based on
+   `benches.yml` and CLI arguments, runs the strategies on dataset questions, and saves the raw LLM outputs and timings to
+   a JSONL file.
+2. **`benches/evaluate.py`**: Handles the evaluation phase. It reads the results file generated by `run.py`, extracts
+   final answers from the raw outputs (using either heuristics or an LLM), compares them to gold answers, and calculates
+   accuracy metrics for each method.
 
 ## Configuration
 
@@ -23,7 +23,7 @@ command-line arguments passed to the scripts.
 
 Detailed instructions on how to configure and run the benchmarks are available in the benchmarking README:
 
-[**View Benchmarking README (`benches/README.md`)**](https://github.com/habedi/cogitator/blob/main/benches/README.md)
+[View Benchmarking README (`benches/README.md`)](https://github.com/habedi/cogitator/blob/main/benches/README.md)
 
 This README covers:
