+++ docs/contributing.md
@@ -3,4 +3,4 @@
 Contributions to Cogitator are welcome! Please refer to the main contribution guidelines document in the repository root
 for details on how to contribute, report bugs, suggest features, and the development workflow.
 
-[**View Contribution Guidelines (`CONTRIBUTING.md`)**](https://github.com/habedi/cogitator/blob/main/CONTRIBUTING.md)
+[Contribution Guidelines (`CONTRIBUTING.md`)](https://github.com/habedi/cogitator/blob/main/CONTRIBUTING.md)
