+++ examples/README.md
@@ -1,15 +1,15 @@
 ## Examples
 
-| File                                                 | Description                                                           |
-|------------------------------------------------------|-----------------------------------------------------------------------|
-| [run_simple_example.py](run_simple_example.py)       | A simple end-to-end example of using the Cogitator library            |
-| [run_least_to_most.py](run_least_to_most.py)         | Example of using the Least-to-Most prompting method                   |
-| [run_sc_cot.py](run_sc_cot.py)                       | Example of using the Self-Consistency prompting method                |
-| [run_auto_cot.py](run_auto_cot.py)                   | Example of using the Automatic CoT prompting method                   |
-| [run_tree_of_thoughts.py](run_tree_of_thoughts.py)   | Example of using the Tree of Thoughts prompting method                |
-| [run_graph_of_thoughts.py](run_graph_of_thoughts.py) | Example of using the Graph of Thoughts prompting method               |
-| [run_cdw_cot.py](run_cdw_cot.py)                     | Example of using the Clustered Distance-Weighted CoT prompting method |
-| [shared.py](shared.py)                               | Shared utilies and settings for the examples                          |
+| File                                                 | Description                                                              |
+|------------------------------------------------------|--------------------------------------------------------------------------|
+| [run_simple_example.py](run_simple_example.py)       | A simple end-to-end example of using the Cogitator library               |
+| [run_least_to_most.py](run_least_to_most.py)         | Example of using the Least-to-Most prompting strategy                    |
+| [run_sc_cot.py](run_sc_cot.py)                       | Example of using the Self-Consistency prompting strategy                 |
+| [run_auto_cot.py](run_auto_cot.py)                   | Example of using the Automatic CoT prompting strategy                    |
+| [run_tree_of_thoughts.py](run_tree_of_thoughts.py)   | Example of using the Tree of Thoughts prompting framework                |
+| [run_graph_of_thoughts.py](run_graph_of_thoughts.py) | Example of using the Graph of Thoughts prompting framework               |
+| [run_cdw_cot.py](run_cdw_cot.py)                     | Example of using the Clustered Distance-Weighted CoT prompting framework |
+| [shared.py](shared.py)                               | Shared utilies and settings for the examples                             |
 
 ## Running Examples
 
@@ -34,5 +34,4 @@ make example-openai
 ```
 
 Note that the examples should be run from the root directory of the repository.
-Additionally, to use `gemma3:4b` (or any other model like `gemma3:12b`) with Ollama, it must be pulled (or downloaded)
-first.
+Additionally, to use `gemma3:4b` (or any other model like `gemma3:12b`) with Ollama, it must be pulled (or downloaded) first.
