+++ examples/run_simple_example.py
@@ -15,9 +15,9 @@ except Exception as e:
     print("Please make sure Ollama is running and the model is pulled.")
     exit(1)
 
-# Step 3: Choose a CoT method (Self-Consistency in this case)
+# Step 3: Choose a CoT strategies (Self-Consistency in this case)
 # Self-Consistency generates multiple reasoning paths and finds the most common answer
-sc_method = SelfConsistency(
+sc_strategy = SelfConsistency(
     llm,
     n_samples=5,  # Number of reasoning paths to generate
     temperature=0.7  # Higher temperature can lead to more diverse answers
@@ -27,10 +27,10 @@ sc_method = SelfConsistency(
 question = "A bat and a ball cost $1.10 in total. The bat costs $1.00 more than the ball. How much does the ball cost?"
 prompt = f"Q: {question}\nA: Let's think step by step."
 
-# Step 5: Run the CoT prompting method
+# Step 5: Run the CoT prompting sc_strategy
 print(f"\nQuestion: {question}")
 print("Running Self-Consistency CoT...")
-final_answer = sc_method.run(prompt)  # Returns the most consistent (repeated) answer
+final_answer = sc_strategy.run(prompt)  # Returns the most consistent (repeated) answer
 
 # Expected output: $0.05 or 0.05 (may vary slightly based on model and temperature)
 print(f"\nCogitator's Answer (Self-Consistency): {final_answer}")
