+++ examples/run_tree_of_thoughts.py
@@ -20,7 +20,7 @@ QUESTIONS = [
 ]
 
 
-async def main_async(args: argparse.Namespace) -> None:
+async def main_async(args: argparse.Namespace):
     llm = get_llm(args.provider, args.model_name, args.openai_key)
     tot = setup_tot(llm)
     semaphore = asyncio.Semaphore(5)
@@ -29,17 +29,18 @@ async def main_async(args: argparse.Namespace) -> None:
     tasks = [tot.run_async(q, semaphore=semaphore) for q in QUESTIONS]
     answers = await asyncio.gather(*tasks)
 
-    for _q, _a in zip(QUESTIONS, answers, strict=False):
-        pass
+    for q, a in zip(QUESTIONS, answers):
+        print(f"Q: {q}\nA: {a}\n")
 
 
-def main_sync(args: argparse.Namespace) -> None:
+def main_sync(args: argparse.Namespace):
     llm = get_llm(args.provider, args.model_name, args.openai_key)
     tot = setup_tot(llm)
 
     logger.info("Running TreeOfThoughts synchronously...")
     for q in QUESTIONS:
-        tot.run(q)
+        a = tot.run(q)
+        print(f"Q: {q}\nA: {a}\n")
 
 
 if __name__ == "__main__":
