+++ pyproject.toml
@@ -127,7 +127,7 @@ target-version = "py310"
 unsafe-fixes = true
 
 [tool.ruff.lint]
-select = ["ANN", "E", "F", "I", "W", "B", "RUF", "SIM", "C90", "T20"]
+select = ["ANN", "E", "F", "I", "W", "B", "RUF", "SIM", "C90"]
 ignore = [
     # Ignore docstring errors
     "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107",
